add scripts for DB's constraints and pagination for Avatar table

+++ src/main/java/com/evgeniyfedorchenko/hogwarts/services/StudentServiceImpl.java
@@ -105,7 +105,7 @@ public class StudentServiceImpl implements StudentService {
     }
 
     @Override
-    public Integer getAverageAge() {
+    public Double getAverageAge() {
         return studentRepository.getAverageAge();
     }
